apps/producer/requirements.txt
kafka-python
requests
aiokafka
trino
flask